requirements.txt
playwright==1.54.0
beautifulsoup4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
